fix(permissions):補上頁尾標籤模組icon權限顯示及調整新聞分類/合作夥伴/廣告管理的權限icon設定

+++ public/index.php
@@ -80,19 +80,16 @@ switch($page) {
     // 後台：文章管理
     case 'article_index':
         requirePermission('view', MODULE_ARTICLE);
-
         $controller = new ArticleController();
         $controller->index();
         break;
     case 'article_create':
         requirePermission('create', MODULE_ARTICLE);
-
         $controller = new ArticleController();
         $controller->create();
         break;
     case 'article_store':
         requirePermission('create', MODULE_ARTICLE);
-
         if($_SERVER['REQUEST_METHOD'] === 'POST'){
             $controller = new ArticleController();
             $controller->store();
@@ -102,13 +99,11 @@ switch($page) {
         break;
     case 'article_edit':
         requirePermission('edit', MODULE_ARTICLE);
-
         $controller = new ArticleController();
         $controller->edit($_GET['id'] ?? null);
         break;
     case 'article_update':
         requirePermission('edit', MODULE_ARTICLE);
-
         if($_SERVER['REQUEST_METHOD'] === 'POST'){
         $controller = new ArticleController();
         $controller->update($_POST);
@@ -116,7 +111,6 @@ switch($page) {
         break;
     case 'article_delete':
         requirePermission('delete', MODULE_ARTICLE);
-
         $controller = new ArticleController();
         $controller->delete($_GET['id'] ?? null);
         break;
@@ -131,13 +125,11 @@ switch($page) {
     // 後台：新聞分類管理
     case 'category_index':
         requirePermission('view', MODULE_CATEGORY);
-
         $controller = new NewsCategoryController();
         $controller->index();
         break;
     case 'category_store':
         requirePermission('create', MODULE_CATEGORY);
-
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $controller = new NewsCategoryController();
             $controller->store();
@@ -145,16 +137,15 @@ switch($page) {
             echo "請使用表單送出資料";
         }
         break;
+
     // 後台：廣告管理
     case 'sponsorpicks_index':
         requirePermission('view', MODULE_SPONSORED);
-
         $controller = new SponsorPickController();
         $controller->index();
         break;
     case 'sponsorpicks_store':
         requirePermission('create', MODULE_SPONSORED);
-
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $controller = new SponsorPickController();
             $controller->store();
@@ -183,36 +174,36 @@ switch($page) {
 
     // 後台：頁尾標籤管理
     case 'footer_index':
-        // requirePermission('create', MODULE_FOOTER);
+        requirePermission('view', MODULE_FOOTER);
         $controller = new footerArticleController();
         $controller->index();
         break;
     case 'footer_create':
-        // requirePermission('create', MODULE_FOOTER);
+        requirePermission('create', MODULE_FOOTER);
         $controller = new footerArticleController();
         $controller->create();
         break;
     case 'footer_store':
-        // requirePermission('create', MODULE_FOOTER);
+        requirePermission('create', MODULE_FOOTER);
         $controller = new footerArticleController();
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         $controller->store();
         }
         break;
     case 'footer_edit':
-        // requirePermission('edit', MODULE_FOOTER);
+        requirePermission('edit', MODULE_FOOTER);
         $controller = new footerArticleController();
         $controller->edit($_GET['id'] ?? null);
         break;
     case 'footer_update':
-        // requirePermission('create', MODULE_FOOTER);
+        requirePermission('edit', MODULE_FOOTER);
         $controller = new footerArticleController();
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         $controller->update($_POST);
         }
         break;
     case 'footer_delete':
-        // requirePermission('create', MODULE_FOOTER);
+        requirePermission('delete', MODULE_FOOTER);
         $controller = new footerArticleController();
         $controller->delete($_GET['id'] ?? null);
         break;
@@ -223,6 +214,7 @@ switch($page) {
         break;
     // 處理footer排序
     case 'api_footer_sort':
+        requirePermission('create', MODULE_FOOTER);
         $controller = new footerArticleController();
         $controller->updateSort();
         break;
@@ -263,19 +255,16 @@ switch($page) {
     // 最高管理者帳號管理(SA)
     case 'sysuser_list':
         requirePermission('view', MODULE_SYSUSER);
-
         $controller = new SysuserController();
         $controller->index();
         break;
     case 'sysuser_create':
         requirePermission('create', MODULE_SYSUSER);
-
         $controller = new SysuserController();
         $controller->create();
         break;
     case 'sysuser_store':
         requirePermission('create', MODULE_SYSUSER);
-
         if($_SERVER['REQUEST_METHOD'] === 'POST'){
         $controller = new SysuserController();
         $controller->store();
@@ -283,13 +272,11 @@ switch($page) {
         break;
     case 'sysuser_edit':
         requirePermission('edit', MODULE_SYSUSER);
-
         $controller = new SysuserController();
         $controller->edit($_GET['id'] ?? null);
         break;
     case 'sysuser_update':
         requirePermission('edit', MODULE_SYSUSER);
-
         if($_SERVER['REQUEST_METHOD'] === 'POST'){
         $controller = new SysuserController();
         $controller->update($_POST);
@@ -297,17 +284,17 @@ switch($page) {
         break;
     case 'sysuser_toggle_status':
         $controller = new SysuserController();
-        $controller->toggle($_GET['id'] ?? null);
+        // $controller->toggle($_GET['id'] ?? null);
         break;
     case 'sysuser_delete':
         requirePermission('delete', MODULE_SYSUSER);
-
         if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
             abort403('非法請求方式');
         }
         $controller = new SysuserController();
         $controller->delete($_POST['id'] ?? null);
         break;
+        
     // 一般管理者profile
     case 'profile':
         $controller = new ProfileController();
